# plugins/llm-service/openapi.yaml
openapi: 3.1.0
info:
  title: LLM Service API
  description: |
    A production-ready FastAPI wrapper around vLLM for on-premise LLM inference.
    
    ## Features
    - Text generation with streaming
    - Chat completion
    - Text classification
    - Entity extraction
    - Text summarization
    - Embeddings generation
    - **Vision OCR** - Intelligent text extraction from images
    - **Data Transform** - LLM-powered data transformation
    
    ## Authentication
    API key authentication via `X-API-Key` header (when enabled).
    
  version: 1.1.0
  contact:
    name: FlowForge Team
  license:
    name: MIT
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.flowforge.io/llm
    description: Production server
...
paths:
...
  /api/v1/vision/ocr:
    post:
      tags: [Vision]
      summary: Vision OCR
      description: |
        Extract text from images using vision-language models.
        Unlike traditional OCR, vision-based OCR understands context,
        handles handwriting, and can preserve layout.
      operationId: visionOCR
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VisionOCRRequest'
      responses:
        '200':
          description: Successful OCR
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VisionOCRResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /api/v1/vision/describe:
    post:
      tags: [Vision]
      summary: Describe image
      description: Generate a natural language description of an image.
      operationId: visionDescribe
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VisionDescribeRequest'
      responses:
        '200':
          description: Successful description
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VisionDescribeResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/v1/vision/extract:
    post:
      tags: [Vision]
      summary: Extract structured data from image
      description: Extract specific fields from documents, forms, receipts, invoices, etc.
      operationId: visionExtract
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VisionExtractRequest'
      responses:
        '200':
          description: Successful extraction
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VisionExtractResponse'
        '400':
          $ref: '#/components/responses/BadRequest'

  # Transform Endpoints
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          default: false
        error:
          type: string
        code:
          type: string
        details:
          type: object

    # Vision Schemas
    ImageInput:
      type: object
      properties:
        url:
          type: string
          description: URL to the image
        base64:
          type: string
          description: Base64-encoded image data
        media_type:
          type: string
          default: image/png
    VisionOCRRequest:
      type: object
      required:
        - image
      properties:
        image:
          $ref: '#/components/schemas/ImageInput'
        language_hints:
          type: array
          items:
            type: string
        output_format:
          type: string
          enum: [text, structured, markdown]
          default: text
        detail_level:
          type: string
          enum: [low, medium, high]
          default: medium
        preserve_layout:
          type: boolean
          default: false
        model:
          type: string
    VisionOCRResponse:
      type: object
      properties:
        success:
          type: boolean
        text:
          type: string
        regions:
          type: array
          items:
            type: object
            properties:
              text:
                type: string
              confidence:
                type: number
              bbox:
                type: array
                items:
                  type: integer
        confidence:
          type: number
        word_count:
          type: integer
        model:
          type: string
        processing_time_ms:
          type: integer
    VisionDescribeRequest:
      type: object
      required:
        - image
      properties:
        image:
          $ref: '#/components/schemas/ImageInput'
        prompt:
          type: string
        max_length:
          type: integer
          default: 200
        focus:
          type: string
          enum: [general, objects, text, people, scene]
          default: general
        model:
          type: string
    VisionDescribeResponse:
      type: object
      properties:
        success:
          type: boolean
        description:
          type: string
        objects_detected:
          type: array
          items:
            type: string
        text_detected:
          type: string
        model:
          type: string
    VisionExtractRequest:
      type: object
      required:
        - image
        - fields
      properties:
        image:
          $ref: '#/components/schemas/ImageInput'
        fields:
          type: array
          items:
            type: string
          minItems: 1
        document_type:
          type: string
        schema:
          type: object
        examples:
          type: array
          items:
            type: object
        model:
          type: string
    VisionExtractResponse:
      type: object
      properties:
        success:
          type: boolean
        fields:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              value: {}
              confidence:
                type: number
        raw_data:
          type: object
        document_type:
          type: string
        confidence:
          type: number
        model:
          type: string

    # Transform Schemas
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ServiceUnavailable:
      description: Service unavailable
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key